go/runtime: Remove obsolete transaction types
