+++ .changelog/4228.breaking.md
@@ -1,0 +1,5 @@
+go/runtime: Remove obsolete transaction types
+
+The deprecated types `TxnCall`, `TxnOutput` and `TxnCheckResult` have been
+removed. Any users should consider migrating to the Runtime SDK or copying
+the type definitions in case they still rely on them.
